+++ assets/assets.go
@@ -2,6 +2,9 @@ package assets
 
 import "embed"
 
+//go:embed config/*
+var Config embed.FS
+
 //go:embed css/*
 var CSS embed.FS